Challenge 6
For each possible keysize we calculate the hamming distance between
several blocks of that keysize length from the ciphertext. The four
keysizes with the lowest distance are chosen and keys of those sizes
are guessed, one byte at a time. Each key is then used to decode the
message, and the result with the highest score is chosen.

- Implement Hamming distance
- Update three::decode_xor to return the key and the message as Vec<u8>
  instead of String.